Adjustment Made To Css
I took margin 0 auto off of the content 2 container and aligned the elements center using justify content. then i adjusted the height of content 2 to 425px to match the height of the first container. i also adjusted the margin top of the h2 heading to 115px, then i adjusted the margin top of the info heading to 50px. i also adjusted the margin top of the text to 3px