Container Added
I added container3 and gave it a height of 250px and colored it grey. i then noticed that the elements in container 2 moved so i aligned them back with the sneaker image. i then noticed that the images in content 2 also moved so i separated them with a gap of 50px. i also changed the name of some of the containers so its more clear and direct